Updated index, css, and script v1.2
Exercises are color coded and the select exercise button reflects that
Body weight seems to be logging correctly
Progress notes logging correctly

Still having issues trying to get individual exercises to log and continue to update the same chart without creating a new one. Trying to get them to appear on the home page instead of the progress page.