Updated changed
Users can now delete progress notes and refresh and everything works as it should - meaning if there are no notes, the charts are deleted, and if there are notes, the local storage is working and redraws the charts and remembers progress.

Starting to come together! Hopefully next are some styling elements and a better navigation.